Refactor import statements and adjust `loadConfig` usage
Refactored import statements in `handler.ts` files across `/commit`, `/changelog`, and `/init` directories. Changed the `loadConfig` usage by specifying types `<InitOptions, InitArgv>`, `<CommitOptions, CommitArgv>`, and `<ChangelogOptions, ChangelogArgv>` to make its functionality clearer. Updated `COMMIT_PROMPT` import to reflect changes in their original files. Added console.log statements for debugging purposes. Changes intend to increase code readability and maintainability.
